refactor(ledger): collocate the transaction-route route registrars :hammer:
The Huma terminal registrar came out of transaction_route_handler.go and the
Fiber guard-chain registrars came out of routes.go; both now live in
transaction_route_routes.go.

No route, authz tuple, operation ID or path changed.
