+++ components/ledger/internal/adapters/http/in/transaction_route_handler.go
@@ -248,76 +248,3 @@ func (handler *TransactionRouteHandler) DeleteTransactionRouteByID(ctx context.C
 
 	return &DeleteTransactionRouteResponse{}, nil
 }
-
-// RegisterTransactionRouteRoutes registers the five transaction-route
-// operations on the shared Huma API. It is the per-file seam the unified server
-// calls; the auth ("midaz","transaction-routes",verb) + tenant +
-// ParseUUIDPathParameters("transaction_route") middleware chain is attached on the
-// /v1 group (Fiber-level) BEFORE the Huma terminal, not here. Paths are
-// GROUP-RELATIVE (see asset_handler.go's RegisterAssetRoutes header for the
-// /v1 rationale).
-//
-// opSuffix distinguishes the operation IDs one version group publishes from another's (empty
-// for /v1, "V2" for /v2 — see routeOpSuffixV1). The v2 twin is a straight mirror: same handler
-// methods, same paths, same input/output types, differing only in the suffixed operation IDs so
-// the two twins do not collide as a duplicate operationId in the one shared document.
-func RegisterTransactionRouteRoutes(api huma.API, h *TransactionRouteHandler, opSuffix string) {
-	const (
-		listPath = "/organizations/{organization_id}/ledgers/{ledger_id}/transaction-routes"
-		idPath   = listPath + "/{transaction_route_id}"
-		tag      = "Transaction Routes"
-	)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "createTransactionRoute" + opSuffix,
-		Method:      http.MethodPost,
-		Path:        listPath,
-		Summary:     "Create Transaction Route",
-		Tags:        []string{tag},
-		Security:    secTransactionRouteBearer,
-		// Body validated imperatively (http.DecodeAndValidate) — see file header.
-		SkipValidateBody: true,
-		DefaultStatus:    http.StatusCreated,
-	}, h.CreateTransactionRoute)
-	attachTypedRequestBody[mmodel.CreateTransactionRouteInput](api, "createTransactionRoute"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "listTransactionRoutes" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        listPath,
-		Summary:     "Get all Transaction Routes",
-		Tags:        []string{tag},
-		Security:    secTransactionRouteBearer,
-	}, h.GetAllTransactionRoutes)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getTransactionRouteByID" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        idPath,
-		Summary:     "Get Transaction Route by ID",
-		Tags:        []string{tag},
-		Security:    secTransactionRouteBearer,
-	}, h.GetTransactionRouteByID)
-
-	huma.Register(api, huma.Operation{
-		OperationID:      "updateTransactionRoute" + opSuffix,
-		Method:           http.MethodPatch,
-		Path:             idPath,
-		Summary:          "Update Transaction Route",
-		Tags:             []string{tag},
-		Security:         secTransactionRouteBearer,
-		SkipValidateBody: true, // body validated imperatively — see file header.
-	}, h.UpdateTransactionRoute)
-	attachTypedRequestBody[mmodel.UpdateTransactionRouteInput](api, "updateTransactionRoute"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "deleteTransactionRoute" + opSuffix,
-		Method:      http.MethodDelete,
-		Path:        idPath,
-		Summary:     "Delete Transaction Route by ID",
-		Tags:        []string{tag},
-		Security:    secTransactionRouteBearer,
-		// DefaultStatus 204 + an Out struct with no Body field => bodiless 204.
-		DefaultStatus: http.StatusNoContent,
-	}, h.DeleteTransactionRouteByID)
-}
